pyproject.toml
[project]
name = "transit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy>=2.0.40",
    "tkintermapview==1.29",
    "Pillow>=10.0.0"
]
